Forward autodump waypoints tested on rover

+++ src/lx_autonomy/src/lx_packages/lx_operation/src/operations_handler.cpp
@@ -453,7 +453,7 @@ bool OperationsHandler::callAutoNav(Task current_task){
     autonav_request_msg.goal.header.stamp = this->get_clock()->now();
     autonav_request_msg.goal.pose = current_task.getPoseArray().poses[0];
     // Check the next task
-    autonav_request_msg.next = task_queue_.front().getType();
+    autonav_request_msg.next = (int)task_queue_.front().getType();
     
     RCLCPP_INFO(this->get_logger(), "Sending Auto Nav goal");
